requirements.txt
azure-ai-documentintelligence>=1.0.0
python-dotenv>=1.0.0
